+++ docs/source/changelog.rst
@@ -1,10 +1,18 @@
-.. Created by changelog.py at 2024-05-24, command
+.. Created by changelog.py at 2024-10-16, command
    '/Users/giffler/.cache/pre-commit/repoecmh3ah8/py_env-python3.12/bin/changelog docs/source/changes compile --categories Added Changed Fixed Security Deprecated --output=docs/source/changelog.rst'
    based on the format of 'https://keepachangelog.com/'
 
 #########
 CHANGELOG
 #########
+
+[Unreleased] - 2024-10-16
+=========================
+
+Added
+-----
+
+* Add support for submit options to HTCondor site adapter
 
 [0.8.2] - 2024-05-15
 ====================